Refactor libaec feature - remove libaec crate and use existing libaec-sys crate
* use existing libaec-sys crate (https://crates.io/crates/libaec-sys)
* move logic from libaec crate directly into ccsds_libaec.rs
