--- libaec-sys/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "libaec-sys"
-version = "0.1.0"
-edition = "2021"
-authors = []
-description = "Raw FFI bindings to libaec (Adaptive Entropy Coding library)"
-license = "MIT"
-repository = "https://github.com/mpiannucci/gribberish"
-
-[build-dependencies]
-bindgen = "0.70"
-cmake = "0.1"
